+++ src/types/OriginDataType/Address.ts
@@ -1,0 +1,5 @@
+export interface Address {
+  depth1: string;
+  depth2: string;
+  depth3: string;
+}
