+++ src/types/OriginDataType/Pin.ts
@@ -1,0 +1,7 @@
+import { GeoJSONPoint } from "./GeoJSON";
+
+export interface Pin {
+  point: GeoJSONPoint;
+  content: string;
+  thumbnailUrl: string | null;
+}
