+++ src/types/OriginDataType/index.ts
@@ -1,13 +1,3 @@
-import { GeoJSONPoint } from "./GeoJSON";
-
-export interface Address {
-  depth1: string;
-  depth2: string;
-  depth3: string;
-}
-
-export interface Pin {
-  point: GeoJSONPoint;
-  content: string;
-  thumbnailUrl: string | null;
-}
+export * from "./Address";
+export * from "./Pin";
+export * from "./GeoJSON";
