+++ src/types/Request/index.ts
@@ -1,0 +1,1 @@
+export * from "./Masils";
